[DSL] Refactor Utils
