+++ src/main/scala/utils/FileUtils.scala
@@ -2,6 +2,8 @@ package utils
 
 import java.io.{File, PrintWriter}
 
+import scala.io.Source
+
 /** Helper functions useful for manipulating files */
 object FileUtils {
 
@@ -10,5 +12,12 @@ object FileUtils {
     val writer = new PrintWriter(new File(filename))
     try writer.write(s)
     finally writer.close()
+  }
+
+  /** Read a string from a file */
+  def readFile(filename: String): String = {
+    val file = Source.fromFile(name = filename)
+    try file.mkString
+    finally file.close()
   }
 }
